+++ src/gamspy/_container.py
@@ -16,6 +16,7 @@ import gamspy as gp
 import gamspy._miro as miro
 import gamspy.utils as utils
 from gamspy._backend.backend import backend_factory
+from gamspy._extrinsic import ExtrinsicLibrary
 from gamspy._miro import MiroJSONEncoder
 from gamspy._options import Options, _map_options
 from gamspy._symbols.symbol import Symbol
@@ -417,7 +418,7 @@ class Container(gt.Container):
             elif isinstance(statement, gp.UniverseAlias):
                 continue
             else:
-                string += statement.getStatement() + "\n"
+                string += statement.getDeclaration() + "\n"
 
         for symbol_name in modified_names:
             symbol = self[symbol_name]
@@ -1041,6 +1042,37 @@ class Container(gt.Container):
 
         if user_invoked:
             self._run()
+
+    def importExtrinsicLibrary(
+        self, lib_path: str, functions: dict[str, str]
+    ) -> ExtrinsicLibrary:
+        """
+        Imports an extrinsic library to the GAMS environment.
+
+        Parameters
+        ----------
+        lib_path : str
+            Path to the .so, .dylib or .dll file that contains the extrinsic library
+        functions : dict[str, str]
+            Names of the functions as a dictionary. Key is the desired function name in GAMSPy
+            and value is the function name in the extrinsic library.
+
+        Returns
+        -------
+        ExtrinsicLibrary
+
+        Raises
+        ------
+        FileNotFoundError
+            In case the extrinsic library does not exist in the given path.
+        """
+        if not os.path.exists(lib_path):
+            raise FileNotFoundError(f"`{lib_path}` is not a valid path.")
+
+        external_lib = ExtrinsicLibrary(lib_path, functions)
+        self._add_statement(external_lib)
+
+        return external_lib
 
     def gamsJobName(self) -> str | None:
         """