+++ CHANGELOG.md
@@ -5,8 +5,10 @@ GAMSPy 0.12.5
 -------------
 - General
   - Do not pick the default solver if the given solver is not compatible with the problem type.
+  - Add extrinsic function support.
   - Expose addGamsCode to user.
 - Testing
+  - Add tests for extrinsic functions.
   - Test whether the given solver is capable of solving the problem type.
   - Add an addGamsCode test for each problem type. 
   - Test Jupyter Notebooks in docs automatically.